Completes refactoring of SearchDevModule
