+++ kifi-backend/modules/search/app/com/keepit/search/ResultClickTracker.scala
@@ -53,7 +53,7 @@ class ResultClickTracker(lru: ProbablisticLRU) {
 
 trait ResultFeedbackModule extends ScalaModule
 
-case class ResultFeedbackImplModule() extends ResultFeedbackModule {
+case class ProdResultFeedbackModule() extends ResultFeedbackModule {
 
   def configure() {}
 
@@ -73,3 +73,28 @@ case class ResultFeedbackImplModule() extends ResultFeedbackModule {
     ResultClickTracker(dir, numHashFuncs, syncEvery)
   }
 }
+
+case class DevResultFeedbackModule() extends ResultFeedbackModule {
+
+  def configure() {}
+
+  @Provides
+  @Singleton
+  def resultClickTracker: ResultClickTracker = {
+    val conf = current.configuration.getConfig("result-click-tracker").get
+    val numHashFuncs = conf.getInt("numHashFuncs").get
+    val syncEvery = conf.getInt("syncEvery").get
+    conf.getString("dir") match {
+      case None => ResultClickTracker(numHashFuncs)
+      case Some(dirPath) =>
+        val dir = new File(dirPath).getCanonicalFile()
+        if (!dir.exists()) {
+          if (!dir.mkdirs()) {
+            throw new Exception("could not create dir %s".format(dir))
+          }
+        }
+        ResultClickTracker(dir, numHashFuncs, syncEvery)
+    }
+  }
+}
+
